Diploma: part 3, test fix
