+++ src/test/java/ru/praktikum/stellarburgers/test/ConstructorTest.java
@@ -1,34 +1,29 @@
 package ru.praktikum.stellarburgers.test;
 
+import org.junit.After;
 import org.junit.Test;
-import org.openqa.selenium.By;
-import ru.praktikum.stellarburgers.ConstructorElements;
-
-import static com.codeborne.selenide.Condition.text;
-import static com.codeborne.selenide.Selenide.$;
+import ru.praktikum.stellarburgers.pageobjects.ConstructorPage;
 import static com.codeborne.selenide.Selenide.open;
 
 public class ConstructorTest {
 
-    ConstructorElements constructorElements =
+    ConstructorPage constructorPage =
             open("https://stellarburgers.nomoreparties.site",
-                    ConstructorElements.class);
+                    ConstructorPage.class);
 
     //переход к соусам
     @Test
     public void toDressings() {
-    constructorElements.setDressings();
-
-    $(By.xpath("//div[@id='root']/div/main/section/div[2]/ul[2]/a[4]/p")).shouldHave(text("Соус с шипами Антарианского плоскоходца"));
+    constructorPage.setDressings();
+    constructorPage.getDressings("Соус с шипами Антарианского плоскоходца");
 
     }
 
     //переход к начиникам
     @Test
     public void toFillings() {
-    constructorElements.setFillings();
-
-    $(By.xpath("//div[@id='root']/div/main/section/div[2]/ul[3]/a[3]/p")).shouldHave(text("Биокотлета из марсианской"));
+    constructorPage.setFillings();
+    constructorPage.getFillings("Биокотлета из марсианской");
 
     }
 
@@ -36,14 +31,13 @@ public class ConstructorTest {
     @Test
     public void toRolls() {
     //чтобы проверить переход к булкам сначала переместимся к другому меню
-    constructorElements.setFillings();
-    //теперь к булкам
-    constructorElements.setRolls();
+    constructorPage.setFillings();
 
-    $(By.xpath("//div[@id='root']/div/main/section/div[2]/ul/a")).shouldHave(text("Флюоресцентная булка R2-D3"));
+    //теперь к булкам
+    constructorPage.setRolls();
+    constructorPage.getRolls("Флюоресцентная булка R2-D3");
 
     }
-
 
 
 }
